+++ cookbooks/yum-nativex/metadata.rb
@@ -1,8 +1,8 @@
-name             'nativex-yum'
+name             'yum-nativex'
 maintainer       'NativeX'
 maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
-description      'Installs/Configures nativex-yum'
+description      'Installs/Configures yum-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'