+++ cookbooks/yum-nativex/recipes/customrepo.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-yum
+# Cookbook Name:: yum-nativex
 # Recipe:: customrepo.rb
 #
 # Copyright 2014, NativeX
@@ -9,7 +9,7 @@
 
 yum_repository 'nativex' do
   description 'NativeX Custom repository'
-  baseurl 'http://spacewalk.teamfreeze.com/nativex-custom'
+  baseurl 'http://spacewalk.teamfreeze.com/custom-nativex'
   gpgcheck false
   gpgkey 'http://spacewalk.teamfreeze.com/pub/rpm-gpg-keys/RPM-GPG-KEY-NativeX-Custom'
   action :create