+++ cookbooks/yum-nativex/recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-yum
+# Cookbook Name:: yum-nativex
 # Recipe:: default
 #
 # Copyright 2014, NativeX 
@@ -7,6 +7,6 @@
 # All rights reserved - Do Not Redistribute
 #
 
-include_recipe "nativex-yum::customrepo"
-include_recipe "nativex-yum::doupgrade"
-include_recipe "nativex-yum::deleterepo"
+include_recipe "yum-nativex::customrepo"
+include_recipe "yum-nativex::doupgrade"
+include_recipe "yum-nativex::deleterepo"