+++ cookbooks/yum-nativex/CHANGELOG.md
@@ -1,11 +1,11 @@
-nativex-yum CHANGELOG
+yum-nativex CHANGELOG
 =====================
 
-This file is used to list changes made in each version of the nativex-yum cookbook.
+This file is used to list changes made in each version of the yum-nativex cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of nativex-yum
+- [your_name] - Initial release of yum-nativex
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.